Refactor codebase types definition and cleanup

+++ src/routes/(api)/api/delete-cube-from-collection/+server.ts
@@ -1,23 +1,29 @@
 import type { RequestHandler } from "./$types";
 import { json } from "@sveltejs/kit";
 
-export const POST: RequestHandler = async ({ request, locals }) => {
+export const POST: RequestHandler = async ({
+  request,
+  locals: { user, supabase },
+}) => {
+  if (!user)
+    return json({ success: false, error: "Unauthorized" }, { status: 401 });
+
   const {
     slug,
   }: {
     slug: string;
   } = await request.json();
 
-  const { error: err } = await locals.supabase
+  const { error: err } = await supabase
     .from("user_cubes")
     .delete()
-    .eq("user_id", locals.user?.id)
+    .eq("user_id", user.id)
     .eq("cube", slug);
 
   if (err)
     return json(
       { success: false, error: "An error occurred: " + err.message },
-      { status: 500 }
+      { status: 500 },
     );
 
   return json({ success: true });
